Refactor scripts
- Remove function code duplication (move main functions to own file)
- JSDoc for main function file
- Remove globals in main function file (use dependency injection)
