+++ 3_calculateMarkovChain.js
